WIP random wind adding, but normal static wind broken

+++ TheSimulation/TheSimulation/Data/Config/Fire/FireConfig.cs
@@ -2,6 +2,6 @@
 
 public sealed record FireConfig
 (
-    float SpreadChancePercent,
+    double SpreadChancePercent,
     bool PauseDuringFire
 );
