+++ TheSimulation/TheSimulation/Data/Config/Prefill/PrefillConfig.cs
@@ -3,5 +3,5 @@
 public sealed record PrefillConfig
 (
     bool ShouldPrefillMap,
-    float Density // neu: 0..1, z.B. 0.7 = 70% der möglichen Bäume
+    double Density // neu: 0..1, z.B. 0.7 = 70% der möglichen Bäume
 );
