+++ TheSimulation/TheSimulation/Helpers/FormatHelper.cs
@@ -2,9 +2,9 @@
 
 public static class FormatHelper
 {
-    public static string FormatTreeDensityText(int activeTreeCount, int maxTreesPossible)
+    public static string FormatTreeDensityText(int activeTreeCount, double maxTreesPossible)
     {
-        var density = activeTreeCount / (float)maxTreesPossible;
+        var density = activeTreeCount / maxTreesPossible;
         var densityPercent = (int)Math.Round(density * 100);
 
         return $"{activeTreeCount} / {maxTreesPossible} ({densityPercent}%)";
